demo/requirements.txt
langchain-groq
langchain
langchain-core
python-dotenv
requests
